+++ spade-hir/src/lib.rs
@@ -24,7 +24,7 @@ use spade_types::PrimitiveType;
 #[derive(PartialEq, Debug, Clone, Serialize, Deserialize)]
 pub struct Block {
     pub statements: Vec<Loc<Statement>>,
-    pub result: Loc<Expression>,
+    pub result: Option<Loc<Expression>>,
 }
 impl WithLocation for Block {}